+++ CHANGELOG.md
@@ -3,12 +3,17 @@
 
 <!-- CHANGELOG NEWEST -->
 
-- Add `operationName` to graphQL requests [#5418](https://github.com/pubkey/rxdb/pull/5418)
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->
 
 <!-- RELEASE BELOW -->
+
+### 15.1.0 (2 January 2024)
+
+- Add `operationName` to graphQL requests [#5418](https://github.com/pubkey/rxdb/pull/5418)
+- Custom fetch to support axios for the graphQL replication [#5421](https://github.com/pubkey/rxdb/pull/5421)
+- FIX CRDT not working with undefined fields [#5423](https://github.com/pubkey/rxdb/pull/5423)
 
 ### 15.0.0 (20 December 2023)